requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
garminconnect==0.2.31
